% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{impute_AAT}
\alias{impute_AAT}
\title{Impute Alpha-1 antitrypsin (AAT)}
\usage{
impute_AAT(GlycA, FAw3, VLDL.D, HDL3.C, LDL.D, Phe, Leu, ApoB, Alb, Tyr, bOHBut,
  BMI, Ala, L.HDL.TG, Ile, Ace, His, HDL.TG, range_check = TRUE,
  standardised = FALSE)
}
\arguments{
\item{GlycA}{NMR measurement for glycoprotein acetyls. Measurements should
be between 0.869--2.24 mmol/L.}

\item{FAw3}{NMR measurement for Omega-3 fatty acids. Measurements should be
between 0.196--1.18 mmol/L.}

\item{VLDL.D}{NMR measurement for the mean diameter of VLDL particles.
Measurements should be between 33--41 nm.}

\item{HDL3.C}{NMR measurement for cholesterol in the HDL-3 fraction.
Measurements should be between 0.421--0.647 mmol/L.}

\item{LDL.D}{NMR measurement for the mean diameter of LDL particles.
Measurements should be between 23.2--24.1 nm.}

\item{Phe}{NMR measurement for phenylalanine. Measurements should be between
0.0555--0.118 mmol/L.}

\item{Leu}{NMR measurement for leucine. Measurements should be between
0.0295--0.138 mmol/L.}

\item{ApoB}{NMR measurement for Apolipoprotein B. Measurements should be
between 0.365--1.48 g/L.}

\item{Alb}{NMR measurement for albumin. Measurements should have a signal
area ranging between 0.0733--0.101.}

\item{Tyr}{NMR measurement for tyrosine. Measurements should be between
0.0279--0.124 mmol/L.}

\item{bOHBut}{NMR measurement for 3-hydroxybutyrate. Measurements should be
between 0.0331--0.872 mmol/L.}

\item{BMI}{Body Mass Index. Measurements should be between 16.17--47.44
kg/m^2.}

\item{Ala}{NMR measurement for alanine. Measurements should be between
0.274--0.557 mmol/L.}

\item{L.HDL.TG}{NMR measurement for triglycerides within large HDL
particles. Measurements should be between 0.000462--0.0933 mmol/L.}

\item{Ile}{NMR measurement for isoleucine. Measurements should be between
0.024--0.103 mmol/L.}

\item{Ace}{NMR measurement for acetate. Measurements should be between
0.029--1.32 mmol/L.}

\item{His}{NMR measurement for histidine. Measurements should be between
0.0446--0.0913 mmol/L.}

\item{HDL.TG}{NMR measurement for triglycerides within all HDL particles.
Measurements should be between 0.0698--0.315 mmol/L.}

\item{range_check}{logical; if \code{TRUE} discard measurements that
are not in the accepted range of values (see Details). If \code{FALSE},
no checking of input measurements or predicted concentrations will
be performed.}

\item{standardised}{logical; have measurements been standardised
(\emph{i.e.} using the \code{scale} function.)}
}
\value{
A vector of AAT measurements ranging between 0.64--2.58 mg/L or
 measurements standardised to the population if \code{standardised = 'TRUE'}.
}
\description{
Imputes the concentrations of AAT from serum NMR measurements in healthy
population-based samples.
}
\details{
The imputation models will only return a concentration where no input
 measurements were missing, and where all input measurements were within
 their acceptable predefined range of values. These correspond to their
 range of values in the training dataset. Similarly, the imputed measurement
 will be set to missing if the imputation model returns a concentration
 outside the range of values that were found in the training dataset.

 Standardised measurements may also be used by setting \code{standardised = 'TRUE'},
 which is useful for cases where measurements must be adjusted for
 technical effects. In this case, all NMR measurements and BMI should
 be log transformed before scaling. Age and sex must also be standardised.
}
