% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{impute_AGP}
\alias{impute_AGP}
\title{Impute Alpha-1-acid glcyoprotein (AGP)}
\usage{
impute_AGP(GlycA, TotFA, IDL.FC, L.HDL.FC, His, HDL.TG, BMI, S.HDL.FC, S.LDL.TG,
  bOHBut, LA, S.HDL.CE, Lac, S.VLDL.TG, Ace, Cit, SFA, Ala, XXL.VLDL.CE, Glol,
  Age, Crea, Gly, range_check = TRUE, standardised = FALSE)
}
\arguments{
\item{GlycA}{NMR measurement for glycoprotein acetyls. Measurements should
be between 0.869--2.24 mmol/L.}

\item{TotFA}{NMR measurement for total fatty acids. Measurements should be
between 5.58--17.1 mmol/L}

\item{IDL.FC}{NMR measurement for free cholesterol within IDL particles.
Measurements should be between 0.0515--0.368 mmol/L.}

\item{L.HDL.FC}{NMR measurement for free cholesterol within large HDL
particles. Measurements should be between 0.00134--0.208 mmol/L.}

\item{His}{NMR measurement for histidine. Measurements should be between
0.0446--0.0913 mmol/L.}

\item{HDL.TG}{NMR measurement for triglycerides within all HDL particles.
Measurements should be between 0.0698--0.315 mmol/L.}

\item{BMI}{Body Mass Index. Measurements should be between 16.17--47.44
kg/m^2.}

\item{S.HDL.FC}{NMR measurement for free cholesterol within small HDL
particles. Measurements should be between 0.0674--0.168 mmol/L.}

\item{S.LDL.TG}{NMR measurement for triglycerides within small LDL
particles. Measurements should be between 0.00478--0.076 mmol/L.}

\item{bOHBut}{NMR measurement for 3-hydroxybutyrate. Measurements should be
between 0.0331--0.872 mmol/L.}

\item{LA}{NMR measurement for 18:2 linoleic acid. Measurements should be
between 0.449--5.15 mmol/L.}

\item{S.HDL.CE}{NMR measurement for cholesterol esters in small HDL
particles. Measurements should be between 0.061--0.466 mmol/L.}

\item{Lac}{NMR measurement for lactate. Measurements should be between
0.69--4.93 mmol/L.}

\item{S.VLDL.TG}{NMR measurement for triglycerides within small VLDL
particles. Measurements should be between 0.0135--0.76 mmol/L.}

\item{Ace}{NMR measurement for acetate. Measurements should be between
0.029--1.32 mmol/L.}

\item{Cit}{NMR measurement for citrate. Measurements should be between
0.0739--0.186 mmol/L.}

\item{SFA}{NMR measurement for saturated fatty acids. Measurements should
be between 2.07--7.02 mmol/L.}

\item{Ala}{NMR measurement for alanine. Measurements should be between
0.274--0.557 mmol/L.}

\item{XXL.VLDL.CE}{NMR measurement for cholesterol esters within extremely
large VLDL particles. Measurements should be between 0.000000133--0.0169
mmol/L.}

\item{Glol}{NMR measurement for glycerol. Measurements should be between
0.0415--0.312 mmol/L.}

\item{Age}{The study participant's age in years. Measurements should be
between 25--74 years old.}

\item{Crea}{NMR measurement for creatinine. Measurements should be between
0.0182--0.347 mmol/L.}

\item{Gly}{NMR measurement for glycine. Measurements should be between
0.182--0.632 mmol/L.}

\item{range_check}{logical; if \code{TRUE} discard measurements that
are not in the accepted range of values (see Details). If \code{FALSE},
no checking of input measurements or predicted concentrations will
be performed.}

\item{standardised}{logical; have measurements been standardised
(\emph{i.e.} using the \code{scale} function.)}
}
\value{
A vector of AGP measurements ranging between 362--1,880 mg/L or
 measurements standardised to the population if \code{standardised = 'TRUE'}.
}
\description{
Imputes the concentrations of AGP from serum NMR measurements in healthy
population-based samples.
}
\details{
The imputation models will only return a concentration where no input
 measurements were missing, and where all input measurements were within
 their acceptable predefined range of values. These correspond to their
 range of values in the training dataset. Similarly, the imputed measurement
 will be set to missing if the imputation model returns a concentration
 outside the range of values that were found in the training dataset.

 Standardised measurements may also be used by setting \code{standardised = 'TRUE'},
 which is useful for cases where measurements must be adjusted for
 technical effects. In this case, all NMR measurements and BMI should
 be log transformed before scaling. Age and sex must also be standardised.
}
