% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{impute_HP}
\alias{impute_HP}
\title{Impute Haptoglobin (HP)}
\usage{
impute_HP(GlycA, LA, IDL.FC, SM, FAw3, HDL.TG, S.VLDL.CE, Age, Alb, Ile, Cit,
  VLDL.D, Leu, Val, L.VLDL.CE, Pyr, Lac, Gln, M.HDL.FC, XL.HDL.TG, XL.HDL.PL,
  His, Tyr, BMI, L.HDL.TG, PUFA, S.LDL.FC)
}
\arguments{
\item{GlycA}{NMR measurement for glycoprotein acetyls. Measurements should
be between 0.869--2.24 mmol/L.}

\item{LA}{NMR measurement for 18:2 linoleic acid. Measurements should be
between 0.449--5.15 mmol/L.}

\item{IDL.FC}{NMR measurement for free cholesterol within IDL particles.
Measurements should be between 0.0515--0.368 mmol/L.}

\item{SM}{NMR measurement for sphingomyelins. Measurements should be between
0.103--0.733 mmol/L.}

\item{FAw3}{NMR measurement for Omega-3 fatty acids. Measurements should be
between 0.196--1.18 mmol/L.}

\item{HDL.TG}{NMR measurement for triglycerides within all HDL particles.
Measurements should be between 0.0698--0.315 mmol/L.}

\item{S.VLDL.CE}{NMR measurement for cholesterol esters within small VLDL
particles. Measurements should be between 0.000221--0.288 mmol/L.}

\item{Age}{The study participant's age in years. Measurements should be
between 25--74 years old.}

\item{Alb}{NMR measurement for albumin. Measurements should have a signal
area ranging between 0.0733--0.101.}

\item{Ile}{NMR measurement for isoleucine. Measurements should be between
0.024--0.103 mmol/L.}

\item{Cit}{NMR measurement for citrate. Measurements should be between
0.0739--0.186 mmol/L.}

\item{VLDL.D}{NMR measurement for the mean diameter of VLDL particles.
Measurements should be between 33--41 nm.}

\item{Leu}{NMR measurement for leucine. Measurements should be between
0.0295--0.138 mmol/L.}

\item{Val}{NMR measurement for valine. Measurements should be between
0.0728--0.297 mmol/L.}

\item{L.VLDL.CE}{NMR measurement for cholesterol esters within large VLDL
particles. Measurements should be between 0.00000152--0.161 mmol/L.}

\item{Pyr}{NMR measurement for pyruvate. Measurements should be between
0.0512--0.179 mmol/L.}

\item{Lac}{NMR measurement for lactate. Measurements should be between
0.69--4.93 mmol/L.}

\item{Gln}{NMR measurement for glutamine. Measurements should be between
0.331--1.36 mmol/L.}

\item{M.HDL.FC}{NMR measurement for free cholesterol within medium HDL
particles. Measurements should be between 0.0208--0.145 mmol/L.}

\item{XL.HDL.TG}{NMR measurement for triglycerides within very large HDL
particles. Measurements should be between 0.000287--0.055 mmol/L.}

\item{XL.HDL.PL}{NMR measurement for phospholipids within very large HDL
particles. Measurements should be between 0.00176--0.639 mmol/L.}

\item{His}{NMR measurement for histidine. Measurements should be between
0.0446--0.0913 mmol/L.}

\item{Tyr}{NMR measurement for tyrosine. Measurements should be between
0.0279--0.124 mmol/L.}

\item{BMI}{Body Mass Index. Measurements should be between 16.17--47.44
kg/m^2.}

\item{L.HDL.TG}{NMR measurement for triglycerides within large HDL
particles. Measurements should be between 0.000462--0.0993 mmol/L.}

\item{PUFA}{NMR measurement for polyunsaturated fatty acids. Measurements
should be between 2.29--6.62 mmol/L.}

\item{S.LDL.FC}{NMR measurement for free cholesterol within small LDL
particles. Measurements should be between 0.0215--0.133 mmol/L.}
}
\value{
A vector of HP measurements ranging between 0.14--3.95 mg/L.
}
\description{
Imputes the concentrations ofHP from serum NMR measurements in healthy
population-based samples.
}
\details{
The imputation models will only return a concentration where no input
 measurements were missing, and where all input measurements were within
 their acceptable predefined range of values. These correspond to their
 range of values in the training dataset. Similarly, the imputed measurement
 will be set to missing if the imputation model returns a concentration
 outside the range of values that were found in the training dataset.
}
