% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{impute_TF}
\alias{impute_TF}
\title{Impute Transferrin (TF)}
\usage{
impute_TF(GlycA, Sex, Age, S.HDL.FC, Ace, Ala, SFA, His, Gln,
  range_check = TRUE, standardised = FALSE)
}
\arguments{
\item{GlycA}{NMR measurement for glycoprotein acetyls. Measurements should
be between 0.869--2.24 mmol/L.}

\item{Sex}{The study participant's sex. The coding should be 1 for males, 2 for
females.}

\item{Age}{The study participant's age in years. Measurements should be
between 25--74 years old.}

\item{S.HDL.FC}{NMR measurement for free cholesterol within small HDL
particles. Measurements should be between 0.0674--0.168 mmol/L.}

\item{Ace}{NMR measurement for acetate. Measurements should be between
0.029--1.32 mmol/L.}

\item{Ala}{NMR measurement for alanine. Measurements should be between
0.274--0.557 mmol/L.}

\item{SFA}{NMR measurement for saturated fatty acids. Measurements should
be between 2.07--7.02 mmol/L.}

\item{His}{NMR measurement for histidine. Measurements should be between
0.0446--0.0913 mmol/L.}

\item{Gln}{NMR measurement for glutamine. Measurements should be between
0.331--1.36 mmol/L.}

\item{range_check}{logical; if \code{TRUE} discard measurements that
are not in the accepted range of values (see Details). If \code{FALSE},
no checking of input measurements or predicted concentrations will
be performed.}

\item{standardised}{logical; have measurements been standardised
(\emph{i.e.} using the \code{scale} function.)}
}
\value{
A vector of TF measurements ranging between 1.39--4.38 mg/L or
 measurements standardised to the population if \code{standardised = 'TRUE'}.
}
\description{
Imputes the concentrations of AGP from serum NMR measurements in healthy
population-based samples.
}
\details{
The imputation models will only return a concentration where no input
 measurements were missing, and where all input measurements were within
 their acceptable predefined range of values. These correspond to their
 range of values in the training dataset. Similarly, the imputed measurement
 will be set to missing if the imputation model returns a concentration
 outside the range of values that were found in the training dataset.

 Standardised measurements may also be used by setting \code{standardised = 'TRUE'},
 which is useful for cases where measurements must be adjusted for
 technical effects. In this case, all NMR measurements and BMI should
 be log transformed before scaling. Age and sex must also be standardised.
}
